Ahora la aplicación puede generar resultados decimales mas largos

+++ Program.cs
@@ -6,7 +6,7 @@ namespace semana_2__proyecto_de_consola_calculadora
     {
         static void Main(string[] args)
         { // Declare variables and then initialize to zero.
-            int num1 = 0; float num2 = 0;
+            Double num1 = 0; Double num2 = 0;
 
             // Display title as the C# console calculator app.
             Console.WriteLine("Console Calculator in C#\r");
@@ -14,11 +14,11 @@ namespace semana_2__proyecto_de_consola_calculadora
 
             // Ask the user to type the first number.
             Console.WriteLine("Type a number, and then press Enter");
-            num1 = Convert.ToInt32(Console.ReadLine());
+            num1 = Convert.ToDouble(Console.ReadLine());
 
             // Ask the user to type the second number.
             Console.WriteLine("Type another number, and then press Enter");
-            num2 = Convert.ToInt32(Console.ReadLine());
+            num2 = Convert.ToDouble(Console.ReadLine());
 
             // Ask the user to choose an option.
             Console.WriteLine("Choose an option from the following list:");
